Major refactoring, WorldPosition class, correct map generation, progress, block creation and destruction

+++ Assets/Shared/SpaceWrapping/Unwrapper3D.cs
@@ -1,6 +1,6 @@
 namespace Shared.SpaceWrapping {
     public class Unwrapper3D : Unwrapper {
-        protected override void Unwrap(int i, out int x, out int y, out int z) => Wrapper.Unwrap(i, out x, out y, out z);
-        protected override int Wrap(int x, int y, int z) => (int) Wrapper.Wrap(x, y, z);
+        protected override void Unwrap(ulong i, out int x, out int y, out int z) => Wrapper.Unwrap(i, out x, out y, out z);
+        protected override ulong Wrap(int x, int y, int z) => Wrapper.Wrap(x, y, z);
     }
 }
